+++ CHANGELOG.md
@@ -2,5 +2,4 @@
 
 ## CLIENT
 
-- added alert reducer and actions to display alerts to the user
-- added Alert component
+- alert user when network error in search